requirements.txt
backtrader
pandas
matplotlib
scikit-learn
xgboost
torch
python-binance
nbformat